// front/src/app/modules/urls/models/url.ts
export class IURL {
    _id: string = null
    url: string = ''
    text: string = ''
    whatsappNumber: string = ''
    whatsappText: string = ''
    companyName: string = ''
    shortUrl: string = ''
    landingID: string = null
    parameters: any = null
    listID: string = ''
    companyID: string = ''
    longURL: string = ''    
    typeURL:string = ''
    nameExcel: string= ''
    dataExcel: Array<any>
    listName: string = ''
    fieldURL: string = ''
    date: Date = null
    code: string = ''
}